Change recipe item design and feed layout to grid
