+++ lib/pages/home_page/home_widgets/recipe_feed_item.dart
@@ -1,0 +1,71 @@
+import 'package:flilipino_food_app/widget_designs/display_recipe.dart';
+import 'package:flutter/material.dart';
+
+class RecipeFeedItem extends StatelessWidget {
+  const RecipeFeedItem(
+      {super.key,
+      required this.name,
+      required this.imageUrl,
+      required this.ingredients,
+      required this.process});
+
+  final String name;
+  final String imageUrl;
+  final String ingredients;
+  final String process;
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: () {
+        Navigator.of(context).push(
+          MaterialPageRoute(
+            builder: (context) => DisplayRecipe(
+              recipeName: name,
+              recipeIngredients: ingredients,
+              recipeProcess: process,
+              recipeImage: imageUrl,
+            ),
+          ),
+        );
+      },
+      child: AspectRatio(
+        aspectRatio: 4 / 3,
+        child: Stack(
+          children: [
+            Positioned.fill(
+              child: Container(
+                decoration:
+                    BoxDecoration(borderRadius: BorderRadius.circular(8)),
+                clipBehavior: Clip.antiAlias,
+                child: Image.network(
+                  imageUrl,
+                  fit: BoxFit.cover,
+                ),
+              ),
+            ),
+            Positioned.fill(
+              child: Container(
+                padding: const EdgeInsets.all(8.0),
+                decoration: const BoxDecoration(
+                    borderRadius: BorderRadius.all(Radius.circular(8)),
+                    gradient: LinearGradient(
+                        colors: [Colors.transparent, Colors.black],
+                        begin: Alignment.topCenter,
+                        end: Alignment.bottomCenter,
+                        stops: [0.4, 1])),
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.end,
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(name),
+                  ],
+                ),
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
